requirements.txt
django==1.7.1
django-cors-headers==1.0.0
django-debug-toolbar==1.2.2
django-extensions==1.4.9
django-gravatar2==1.1.4
django-mptt==0.6.1
django-fsm==2.2.0
django-suit==0.2.12
django-suit-ckeditor==0.0.2
django-redis-cache==0.13.0

djangorestframework==3.0.2
djangorestframework-camel-case==0.2.0

celery==3.1.17
redis==2.10.3

pyyaml==3.11
psycopg2==2.5.4
pillow==2.6.1
sorl-thumbnail==12.1c

-e git+https://github.com/omab/python-social-auth.git#egg=python-social-auth
